refactor(sse): extrai a família MiniMax de services/usage.ts
Continua a decomposição do god-file services/usage.ts (Tier 2). Move a família
MiniMax completa para services/usage/minimax.ts (346L): inferência de plan-label,
montagem de quota por-janela e o fetcher getMiniMaxUsage que sonda os endpoints
de coding-plan remains, mais MINIMAX_USAGE_CONFIG.

- Depende apenas dos leaves irmãos scalars (toRecord/toNumber/getFieldValue/
  pickFirstNonEmptyString) e quota (UsageQuota/parseResetTime/createQuotaFromUsage)
  — sem acoplamento com o host → provider leaf co-localizado.
- usage.ts reimporta as 12 funções MiniMax (dispatcher usa getMiniMaxUsage; o objeto
  __testing re-expõe as 12 para os testes). getClaudePlanLabel (intercalado, não-
  MiniMax) preservado no host.

Acoplamento tratado: o secret-scanner (check-public-creds) tinha 2 FPs congelados
apontando para usage.ts (a assinatura `getMiniMaxUsage(apiKey: ... "minimax")` parece
credencial). Como a função migrou, atualizei a allowlist KNOWN_LITERAL_CREDS e o teste
6A.8 para services/usage/minimax.ts:213.

Behavior-preserving (byte-idêntico). usage.ts 3147 → 2817.

Validação: typecheck:core, eslint 0/0, check:cycles, check:file-size,
check:public-creds OK + teste 6A.8 (19/19), usage-utils (47/47),
minimax-coding-plan-usage (5/5), usage-service-hardening (24/24) + caracterização
nova (4/4, surface + identidade do re-export __testing).
